Remove auto build number code
Removed the build number code that was based on the difference between a
tagged commit and current commit. You can now just pass in a build
number using the -bn or -build-num option.

Also fixed some checks on $help and $use_git_info so that the behaviour is
what is expected.